(PC-12976)[API] refactor: move beneficiary_import_queries to DMS api
